AP-2251 Address review remarks
